% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{addGates}
\alias{addGates}
\title{Add Polygon Gates}
\usage{
addGates(
  gs,
  gateStrat = ".",
  foN.gateStrat = ".",
  type.gateStrat = ".",
  verbose = "."
)
}
\arguments{
\item{gs}{A gating set as produced by \code{\link{makeGatingSet}}.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
A gating set.
}
\description{
Load the predefined gating strategy (as .csv or .xlsx file) and 
apply the gates as defined in the file.
}
\details{
The xlsx file can hold one or more gates. One row in the xls files represents The ending `.xlsx`will be added to the character supplied in
\code{gateStrat} automatically.
}
\seealso{
Other Extraction functions: 
\code{\link{makeGatingSet}()}
}
\concept{Extraction functions}
