% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_accessories.R
\name{applyBandpass}
\alias{applyBandpass}
\title{Apply Bandpass Filter to Fluorescence Intensities}
\usage{
applyBandpass(fdmat, bandpass, gate = 1)
}
\arguments{
\item{fdmat}{An object of \code{\link{class-fdmat}} as produced by
\code{\link{makefdmat}}.}

\item{bandpass}{Numeric vector length two, holding the lower and upper
boundaries of the bandpass filter.}

\item{gate}{Numeric or character length one, defining the gate on which the
bandpass should be applied. Defaults to 1. (See 'fdmat@metadata' for possible
values for gate name resp. number.)}
}
\value{
The same object of class 'fdmat' as was provided to the argument
'fdmat', but with modified range of fluorescence intensities in the gate
specified under 'gate'.
}
\description{
Apply a bandpass like filter to the fluorescence intensities
within a single gate.
}
\examples{
td <- tempdir()
check_download_data(td)
exp_home <- paste0(td, "/flowdex_examples")
old_wd <- getwd()
setwd(exp_home)
#
assign("get_settings_from_flowdex_package_root", TRUE, pos=.GlobalEnv)
# only required to make the examples run automatically
# you should not call 'assign' if you run the examples manually
# the effect of setting 'get_settings_from_flowdex_package_root' to TRUE
# is that the file 'flowdex_settings.R' in 'root' of the installed package
# 'flowdex' will be sourced instead of the one in the user-defined location.
#
fdmat <- flowdexit(patt = "GPos_T4")
fdmat_bp <- applyBandpass(fdmat, c(1600, 2400))
#
setwd(old_wd)
}
\seealso{
Other Accessory functions: 
\code{\link{checkRepairFcsFiles}()},
\code{\link{cutFdmatToGate}()},
\code{\link{exportFdmatData}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()},
\code{\link{genfs}()},
\code{\link{repairSID}()},
\code{\link{repairVolumes}()}
}
\concept{Accessory functions}
