% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{check_download_data}
\alias{check_download_data}
\title{Check and Download Example Dataset}
\usage{
check_download_data(where)
}
\arguments{
\item{where}{Character length one. The path where the example dataset should
be looked for.}
}
\description{
Checks if the example dataset is present. If not, the example
dataset is downloaded.
}
\examples{
td <- tempdir()
check_download_data(td)
}
