% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-cyTags}
\alias{class-cyTags}
\title{Class 'cyTags'}
\description{
Contains a data frame holding the cyTags, i.e. the class and
the numerical variables that were possibly extracted from the structured
sample ID of each sample.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A data frame containing the cyTags}
}}

\seealso{
Other Class documentation: 
\code{\link{class-GatingSet_fd}},
\code{\link{class-eventsPV}},
\code{\link{class-fdmat_single}},
\code{\link{class-fdmat}},
\code{\link{class-gatingStrategy_fd}}
}
\concept{Class documentation}
