% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-eventsPV}
\alias{class-eventsPV}
\title{Class 'eventsPV'}
\description{
Contains a data frame holding the overall events per volume
unit.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A data frame holding the overall events per volume
unit.}

\item{\code{gateName}}{Character. The name of the corresponding gate,}

\item{\code{volumeUnit}}{The character string denoting the volume unit. \cr
The volume unit is, in the end, defined by the volume factor as provided in
the flowdex.settings.R file (key: 'dV_volumeFactor')}
}}

\seealso{
Other Class documentation: 
\code{\link{class-GatingSet_fd}},
\code{\link{class-cyTags}},
\code{\link{class-fdmat_single}},
\code{\link{class-fdmat}},
\code{\link{class-gatingStrategy_fd}}
}
\concept{Class documentation}
