% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-fdmat}
\alias{class-fdmat}
\title{Class 'fdmat'}
\description{
Contains a list, and within each list element there is an object
of class 'fdmat_single', holding in turn the fluorescence distributions. For
each gate in the gating strategy where 'keepData' is set to TRUE there is
one list element in 'fdmat'.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A list with the same length as there are gates, i.e. with the
same length as there are rows in the data frame in the slot \code{metadata}.}

\item{\code{metadata}}{A data frame holding the metadata, i.e. the parameters used
to calculate the fluorescense intensities. Most of these parameters can be set
in the file 'flowdex_settings.R'.}

\item{\code{cyTags}}{An object of \code{\link{class-cyTags}}}

\item{\code{gateStrat}}{An object of \code{\link{class-gatingStrategy_fd}}}

\item{\code{pData}}{A data frame holding the pheno-data, i.e. the data showing the
raw sample-ID and the volume data stored in each single fcs file.}

\item{\code{note}}{Character. Displaying comments like e.g. if
\code{\link{applyBandpass}} was applied to data from a gate.}
}}

\seealso{
Other Class documentation: 
\code{\link{class-GatingSet_fd}},
\code{\link{class-cyTags}},
\code{\link{class-eventsPV}},
\code{\link{class-fdmat_single}},
\code{\link{class-gatingStrategy_fd}}
}
\concept{Class documentation}
