% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-fdmat_single}
\alias{class-fdmat_single}
\title{Class 'fdmat_single'}
\description{
Contains a matrix with the fluorescence distribution from data
coming from a single gate.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A matrix containing the fluorescence distributions}

\item{\code{eventsPerVol}}{An object of \code{\link{class-eventsPV}}}

\item{\code{gateName}}{Character. The name of the gate where the data used to calculate
the fluorescence distributions came from.}

\item{\code{metadata}}{A data frame holding the metadata, i.e. the parameters used
to calculate the fluorescense intensities. Most of these parameters can be set
in the file 'flowdex_settings.R'.}

\item{\code{ncpwl}}{Numeric. The number of characters before the actual fluorescence
intensities in the column names of the matrix at \code{.Data}.}

\item{\code{note}}{Character. Displaying comments like e.g. if
\code{\link{applyBandpass}} was applied to the data.}
}}

\seealso{
Other Class documentation: 
\code{\link{class-GatingSet_fd}},
\code{\link{class-cyTags}},
\code{\link{class-eventsPV}},
\code{\link{class-fdmat}},
\code{\link{class-gatingStrategy_fd}}
}
\concept{Class documentation}
