% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-GatingSet_fd}
\alias{class-GatingSet_fd}
\title{Class 'GatingSet_fd'}
\description{
Is extending the class \code{\link[flowWorkspace]{GatingSet}}.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{An object of class \code{\link[flowWorkspace]{GatingSet}}}

\item{\code{gateStrat}}{An object of  \code{\link{class-gatingStrategy_fd}}}
}}

\seealso{
Other Class documentation: 
\code{\link{class-cyTags}},
\code{\link{class-eventsPV}},
\code{\link{class-fdmat_single}},
\code{\link{class-fdmat}},
\code{\link{class-gatingStrategy_fd}}
}
\concept{Class documentation}
