% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cl_classDefs.R
\docType{class}
\name{class-gatingStrategy_fd}
\alias{class-gatingStrategy_fd}
\title{Class 'gatingStrategy_fd'}
\description{
Contains a data frame with the gating strategy.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{A data frame containing the gating strategy as defined in the
gating strategy file.}

\item{\code{filename}}{The filename of the gating strategy file.}
}}

\seealso{
Other Class documentation: 
\code{\link{class-GatingSet_fd}},
\code{\link{class-cyTags}},
\code{\link{class-eventsPV}},
\code{\link{class-fdmat_single}},
\code{\link{class-fdmat}}
}
\concept{Class documentation}
