% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{cutFdmatToGate}
\alias{cutFdmatToGate}
\title{Cut fdmat to Gate}
\usage{
cutFdmatToGate(fdmat, gate = NULL)
}
\arguments{
\item{fdmat}{An object of class 'fdmat' as produced by \code{\link{makefdmat}}.}

\item{gate}{Numeric or Character length one. The designator for the gate to 
keep, as defined in the gating strategy (from those gates where 'keepData' is 
set to TRUE.}
}
\value{
An object of class `fdmat` containing only the data for the gate as
specified in \code{gate}.
}
\description{
Cut an object of class `fdmat` down to only a single gate
}
\examples{
\dontrun{
gs <- makeAddGatingSet()
fdm <- makefdmat(gs)
fdmCut <- cutFdmatToGate(fdm, 1)
fdmCut <- cutFdmatToGate(fdm, "fooBar")
}
}
\seealso{
Other Helper functions: 
\code{\link{applyBandpass}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()}
}
\concept{Helper functions}
