% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{fd_save}
\alias{fd_save}
\title{Save Fluorescence Distribution 'fdmat' Object}
\usage{
fd_save(fdmat, fns = NULL, expo.folder = ".", verbose = ".")
}
\arguments{
\item{fdmat}{An object of class 'fdmat' as produced by \code{\link{makefdmat}}.}

\item{fns}{Character length one or NULL. Possible character to be added to 
the filename.}

\item{expo.folder}{Character length one. The name of the folder where exported 
should reside. If left at the default '.', the value as defined in the 
settings file (key 'foN_rawData') will be used.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
Invisible NULL; is called for its side effect, i.e. to save the fdmat 
object to file.
}
\description{
Saves the R-object containing the fluorescence distributions 
(the 'fdmat' object) in the standard data export / rawdata folder.
}
\examples{
\dontrun{
fd_save(fdmat)
}
}
\seealso{
Other Helper functions: 
\code{\link{cutFdmatToGate}()},
\code{\link{fd_load}()}
}
\concept{Helper functions}
