% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{fd_update_settings}
\alias{fd_update_settings}
\title{Manually Update the Settings}
\usage{
fd_update_settings(silent = FALSE)
}
\arguments{
\item{silent}{Logical}
}
\value{
A list holding key-value pairs as defined in the 'flowdex_settings.R'
file.
}
\description{
Source the settings-list as defined in the settings.R file and
create the object 'stn' in its environment. The settings-file system is
based on functionality from package \code{\link[uniset]{uniset}}.
}
\examples{
assign("get_settings_from_flowdex_package_root", TRUE, pos=.GlobalEnv)
# only required to make the examples run automatically
# you should not call 'assign' if you run the examples manually
# the effect of setting 'get_settings_from_flowdex_package_root' to TRUE
# is that the file 'flowdex_settings.R' in 'root' of the installed package
# 'flowdex' will be sourced instead of the one in the user-defined location.
#
stn <- fd_update_settings()
str(stn)
}
\seealso{
Other Setup functions: 
\code{\link{setup_settings}()}
}
\concept{Setup functions}
