% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{fd_updateSettings}
\alias{fd_updateSettings}
\title{Manually Update the Settings}
\usage{
fd_updateSettings(silent = FALSE)
}
\arguments{
\item{silent}{Logical}
}
\value{
A list holding key-value pairs as defined in the 'flowdex_settings.R'
file.
}
\description{
Source the settings-list as defined in the settings.R file and
create the object 'stn' in its environment. The settings-file system is
based on functionality from package \code{\link[uniset]{uniset}}.
}
\section{Examples}{
 
Please refer to \url{https://bpollner.github.io/flowdex/} for executable 
examples in the tutorial dataset.
}

\seealso{
\code{\link{setupSettings}}
}
