% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{genfs}
\alias{genfs}
\title{Generate Folder Structure}
\usage{
genfs(copyTemplates = TRUE, where = getwd())
}
\arguments{
\item{copyTemplates}{Logical, if available templates should be copied into 
the folder 'templates'.}

\item{where}{Character length one, holding a valid path. Defaults to the 
current working directory.}
}
\value{
No return value, called for its side effects, i.e. the creation of 
the required folder structure.
}
\description{
Generate the required folder structure, and possibly copy the 
available templates (gate definitions, gating strategy, dictionary).
}
\concept{Helper Functions}
