% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_f_Inner.R
\name{getEventsPerVolume}
\alias{getEventsPerVolume}
\title{Get Events per Volume Unit}
\usage{
getEventsPerVolume(gs)
}
\arguments{
\item{gs}{A gating set as produced by \code{\link{makeAddGatingSet}}.}
}
\value{
A data frame.
}
\description{
From the data contained in the provided gating set, produce a 
data frame containing the events per volume unit for every gate in every 
single flowframe (.e. sample tube).
}
\concept{Statistic functions}
