% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{makeAddGatingSet}
\alias{makeAddGatingSet}
\title{Make Gating Set and Add Gates}
\usage{
makeAddGatingSet(
  patt = NULL,
  fn = ".",
  gateStrat = ".",
  foN.gateStrat = ".",
  type.gateStrat = ".",
  comp = ".",
  tx = ".",
  channel = ".",
  ignore.text.offset = FALSE,
  verbose = "."
)
}
\arguments{
\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file (key: 'foN_fcsFiles') will be used.}

\item{gateStrat}{Character length one. The name of the file defining the 
gating strategy. If left at the default '.', the name as defined in the 
settings file (key: 'fiN_gateStrat') will be used.}

\item{foN.gateStrat}{Character length one. The name of the folder where the 
file defining the gating strategy and the gate definitions reside. If left 
at the default '.', the name as defined in the settings file 
(key: 'foN_gating') will be used.}

\item{type.gateStrat}{Character length one, can be either 'csv' or 'xlsx'. 
files are supported. If left at the default '.', the filetype as defined in 
the settings (key: 'dV_gateStratInputType') file will be used.}

\item{comp}{Logical. If compensation should be applied or not. If left at 
the default '.', the value as defined in the settings file (key 'dV_comp') 
will be used.}

\item{tx}{Character length one. The transformation applied to *all* channels 
within the individual flow sets. If left at the default '.', the value as 
defined in the settings file (key 'dV_tx') will be used. (Currently only 
'fjbiex' is implemented.)}

\item{channel}{A regular expression indicating which channels, i.e. which
columns to keep from the original flowframes; is passed down to argument
\code{column.pattern} of \code{\link[flowCore]{read.flowSet}}. Set to NULL
to read data from all channels. If left at the default '.', the value as 
defined in the settings file (key 'dV_channel') will be used.}

\item{ignore.text.offset}{Logical. If set to true, fcs-files in the folder 
specified at argument 'fn' will be checked for inconsistencies in the 
HEADER and TEXT segment and, if those inconsistency is present, the 
afflicted fcs-file will be re-written to disc, with the values in TEXT 
being ignored. The file will be \strong{overwritten without further warning.}}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
A gating set with added and recomputed gates.
}
\description{
Read in FCS files, put them together in a gating set and apply 
the gating strategy as previously defined.
}
\details{
This is a convenience wrapper for the two functions 
\code{\link{makeGatingSet}} and \code{\link{addGates}}.
}
\examples{
\dontrun{
gs <-  makeAddGatingSet()
plot(gs)
}
}
\seealso{
Other Extraction functions: 
\code{\link{addGates}()},
\code{\link{makeGatingSet}()},
\code{\link{makefdmat}()}
}
\concept{Extraction functions}
