% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{repairVolumes}
\alias{repairVolumes}
\title{Repair Volume Values in FCS files}
\usage{
repairVolumes(
  patt = NULL,
  vol = NULL,
  fn = ".",
  includeAll = FALSE,
  confirm = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.#' @export}

\item{vol}{Numeric length one. The value that should be written into the
\code{$VOL} slot in the description of the single flowFrame.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file will be used.}

\item{includeAll}{Logical. If left at the default \code{FALSE}, only those
FCS files with *missing* volume information are read in and the value provided
in \code{vol} will be assigned to them. If changed to \code{TRUE}, also FCS
files with *present* volume information will be read in and the value provided
in \code{vol} will be assigned to them.}

\item{confirm}{Logical. If the user should be asked for additional confirmation
before the rewriting of the fcs files is performed. Defaults to TRUE.}

\item{verbose}{Logical. If detailed status messages should be displayed.
Defaults to TRUE.}
}
\value{
Nothing, resp. the modified fcs files written back into the same folder
and replacing the original ones.
}
\description{
Read in all or a subset of FCS files, replace missing volume
values in the description of the single flowFrames with a provided default
value and write the FCS-files back to disc.
}
\details{
This function is intended to be used when, for known or unknown
reasons, there is no volume information encoded in the original fcs file.
}
\concept{Repair functions}
