% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{setup_settings}
\alias{setup_settings}
\title{Perform Settings Setup}
\usage{
setup_settings(path = NULL)
}
\arguments{
\item{path}{Character length one, holding the path to the location where the
folder containing the settings.R file should be located. Defaults to 'NULL'.
If left at the default 'NULL', the location should be selectable
interactively.}
}
\value{
(Invisible) NULL; is called for its side effects, i.e. to set up
the settings-file system as provided by package 'uniset'.
}
\description{
Perform the setup of the setting-system provided by package
'uniset'.
}
\details{
Has to be done only once.
}
\examples{
\dontrun{
where <- "~/desktop"
setup_settings(where)
}
}
\seealso{
Other Setup functions: 
\code{\link{fd_update_settings}()}
}
\concept{Setup functions}
