% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{setupSettings}
\alias{setupSettings}
\title{Perform settings setup}
\usage{
setupSettings()
}
\value{
No return value, is called for its side effects, i.e. to set up 
the settings-file system as provided by package 'uniset'.
}
\description{
Perform the setup of the setting-system provided by package 
'uniset'.
}
\details{
Has to be done only once.
}
