% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{cutFdmatToGate}
\alias{cutFdmatToGate}
\title{Cut 'fdmat' Object to Gate}
\usage{
cutFdmatToGate(fdmat, gate = NULL)
}
\arguments{
\item{fdmat}{An object of \code{\link{class-fdmat}} as produced by
\code{\link{makefdmat}}.}

\item{gate}{Numeric or Character length one. The designator for the gate to
keep, as defined in the gating strategy (from those gates where 'keepData' is
set to TRUE.}
}
\value{
An object of \code{\link{class-fdmat}} containing only the data for
the gate as specified in \code{gate}.
}
\description{
Cut an object of class `fdmat` down to only a single gate
}
\section{Examples}{
 
Please refer to \url{https://bpollner.github.io/flowdex/} for executable 
examples in the tutorial dataset.
}

\seealso{
Other Accessory functions: 
\code{\link{applyBandpass}()},
\code{\link{checkRepairFcsFiles}()},
\code{\link{exportFdmatData}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()},
\code{\link{genfs}()},
\code{\link{repairSID}()},
\code{\link{repairVolumes}()}
}
\concept{Accessory functions}
