% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{exportFdmatData}
\alias{exportFdmatData}
\title{Export Fluorescence Distributions}
\usage{
exportFdmatData(
  fdmat,
  expo.gate = ".",
  expo.name = ".",
  expo.type = ".",
  expo.folder = ".",
  verbose = "."
)
}
\arguments{
\item{fdmat}{An object of class 'fdmat' as produced by \code{\link{makefdmat}}.}

\item{expo.gate}{Which gate to export. NULL or numeric or character length 
one. Set to NULL to export data from all those gates defined in the gating 
strategy where 'keeoData' is set to TRUE. Provide a character length one 
with a gate name or the number of that gate as defined in the gating strategy 
to export data from this gate only. If left at the default '.', the value as 
defined in the settings file (key 'dE_exportGate') will be used.}

\item{expo.name}{Character length one. The name of the file holding the 
exported fluorescence distribution(s). If left at the default '.', the value 
as defined in the settings file (key 'fiN_dataExport') will be used.}

\item{expo.type}{Character length one. The filetype of the data export. 
Possible values are 'csv' and 'xlsx'.  If left at the default '.', the value 
as defined in the settings file (key 'dE_exportType') will be used.}

\item{expo.folder}{Character length one. The name of the folder where exported 
should reside. If left at the default '.', the value as defined in the 
settings file (key 'foN_rawData') will be used.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
Invisible NULL; used for its side effects, i.e. to export the data contained in 
'fdmat' to file.
}
\description{
Export fluorescence distributions contained in the 'fdmat' 
object to file.
}
\details{
If data are exported to xlsx, additional data like the metadata 
describing the parameters that lead to the calculation of the fluorescence 
distribution, the cyTags and the gating strategy are saved in an extra sheet 
as well. If exporting to csv, only the fluorescence data are exported.
}
\examples{
\dontrun{
gs <- makeAddGatingSet()
fdm <- makefdmat(gs, expo=FALSE) # export is included here already
exportFdmatData(fdm)
}
}
