% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowdex.R
\docType{package}
\name{flowdex}
\alias{flowdex}
\title{Extract Fluorescence Distribution Data from FCS Files and Recalculate
to Events per Volume}
\description{
Extract fluorescence distribution data from any bivariate
distribution within a previously defined flow cytometry gating set and
recalculate the fluorescence distribution data to events per volume.
To (meaningfully) use flowdex, the FCM files have to contain volumetric
measurement data denoting the acquired sample volume.
}
\details{
When calling \code{\link{flowdexit}}, the gating set "gs" is assigned
to the parent envelope.
}
\section{Important}{
 In order to be able to correctly re-calculate raw events 
to events per volume unit, it is paramount to obtain the correct volume factor 
from the manual of the FCM-machine / the volumetric measurement module. Please
see the section 'Calculating Events per Volume Unit' at 
\code{\link{flowdexit}} for further details.
}

\section{CAVE}{
 Some functions like e.g. \code{\link{repairVolumes}},
\code{\link{repairSID}} or \code{\link{checkRepairFcsFiles}} can re-write
fcs files to disc, thereby overwriting original fcs files. It is therefore 
recommended to work on a copy of the original fcs files.
}

\section{Maintainer}{
 Bernhard Pollner <bernhard.pollner@mac.com>
}

\section{Links}{
 Please see \url{https://bpollner.github.io/flowdex/} for
practical examples and a small tutorial; bug reports can be made at
\url{https://github.com/bpollner/flowdex/issues}.
}

\section{Important functions}{
 \code{\link{flowdexit}}
}

\section{Acknowledgements}{
 This work was made possible by support from IPF
Austria - Georg Huber. \cr
Flow cytometry data were acquired at the Institute
for Hygiene and Medical Microbiology, Medical University of Innsbruck,
Austria.
}

\author{
Bernhard Pollner
}
