% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{flowdexit}
\alias{flowdexit}
\title{Read in FCS Files and Extract Data}
\usage{
flowdexit(fn = ".", patt)
}
\arguments{
\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file will be used.}

\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.#' @export}
}
\description{
XXX
}
