% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{genfs}
\alias{genfs}
\title{Generate Folder Structure}
\usage{
genfs(where = getwd(), copyTemplates = TRUE)
}
\arguments{
\item{where}{Character length one, holding a valid path. Defaults to the
current working directory.}

\item{copyTemplates}{Logical, if available templates should be copied into
the folder 'templates'.}
}
\value{
No return value, called for its side effects, i.e. the creation of
the required folder structure.
}
\description{
Generate the required folder structure, and possibly copy the
available templates (gate definitions, gating strategy, dictionary).
}
\examples{
\dontrun{
home <- paste0(tempdir(), "/home")
dir.create(home)
genfs(home)
}
}
\seealso{
Other Accessory functions: 
\code{\link{applyBandpass}()},
\code{\link{checkRepairFcsFiles}()},
\code{\link{cutFdmatToGate}()},
\code{\link{exportFdmatData}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()},
\code{\link{repairSID}()},
\code{\link{repairVolumes}()}
}
\concept{Accessory functions}
