% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{genfs}
\alias{genfs}
\title{Generate Folder Structure}
\usage{
genfs(where = getwd(), copy_templates = TRUE)
}
\arguments{
\item{where}{Character length one, holding a valid path. Defaults to the
current working directory.}

\item{copy_templates}{Logical, if available templates should be copied into
the folder 'templates'.}
}
\value{
No return value, called for its side effects, i.e. the creation of
the required folder structure.
}
\description{
Generate the required folder structure, and possibly copy the
available templates (gate definitions, gating strategy, dictionary).
}
\examples{
td <- tempdir()
data_source <- "https://github.com/bpollner/data/raw/main/flowdex_examples/flowdex_examples.zip"
check_download_data(td, data_source)
exp_home <- paste0(td, "/flowdex_examples")
old_wd <- getwd()
setwd(exp_home)
#
assign("get_settings_from_flowdex_package_root", TRUE, pos=.GlobalEnv)
# only required to make the examples run automatically
# you should not call 'assign' if you run the examples manually
# the effect of setting 'get_settings_from_flowdex_package_root' to TRUE
# is that the file 'flowdex_settings.R' in 'root' of the installed package
# 'flowdex' will be sourced instead of the one in the user-defined location.
#
genfs(exp_home)
#
setwd(old_wd)
}
\seealso{
Other Accessory functions: 
\code{\link{applyBandpass}()},
\code{\link{checkRepairFcsFiles}()},
\code{\link{cutFdmatToGate}()},
\code{\link{exportFdmatData}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()},
\code{\link{repairSID}()},
\code{\link{repairVolumes}()}
}
\concept{Accessory functions}
