% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_f_Inner.R
\name{getEventsPerVolume}
\alias{getEventsPerVolume}
\title{Get Events per Volume Unit}
\usage{
getEventsPerVolume(gs)
}
\arguments{
\item{gs}{A gating set as produced by \code{\link{makeAddGatingSet}}.}
}
\value{
A list with the same length as there are 'keepData == TRUE' in the
gating strategy file, containing a data frame with the overall events per
volume unit for that gate in each list element.
}
\description{
From the data contained in the provided gating set, obtain a
list containing the events per volume unit in an object of class 'eventsPV'
for every gate in every single flowFrame (.e. sample tube) in a list element.
}
\details{
This function is exported for (internal) back-compatibility. 
Its immediate benefit for the user is rather mild.
}
\examples{
td <- tempdir()
data_source <- "https://github.com/bpollner/data/raw/main/flowdex_examples/flowdex_examples.zip"
check_download_data(td, data_source)
exp_home <- paste0(td, "/flowdex_examples")
old_wd <- getwd()
setwd(exp_home)
#
assign("get_settings_from_flowdex_package_root", TRUE, pos=.GlobalEnv)
# only required to make the examples run automatically
# you should not call 'assign' if you run the examples manually
# the effect of setting 'get_settings_from_flowdex_package_root' to TRUE
# is that the file 'flowdex_settings.R' in 'root' of the installed package
# 'flowdex' will be sourced instead of the one in the user-defined location.
#
gs <- makeAddGatingSet()
evpv <- getEventsPerVolume(gs)
#
setwd(old_wd)
}
\seealso{
\code{\link{flowdexit}}
}
