% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_f_Inner.R
\name{getEventsPerVolume}
\alias{getEventsPerVolume}
\title{Get Events per Volume Unit}
\usage{
getEventsPerVolume(gs)
}
\arguments{
\item{gs}{A gating set as produced by \code{\link{makeAddGatingSet}}.}
}
\value{
A list with the same length as there are 'keepData == TRUE' in the 
gating strategy file.
}
\description{
From the data contained in the provided gating set, obtain a 
list containing the events per volume unit in an object of class 'eventsPV' 
for every gate in every single flowframe (.e. sample tube) in a list element.
}
