% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{makefdmat}
\alias{makefdmat}
\title{Extract Fluorescence Distribution Matrix}
\usage{
makefdmat(
  gs,
  expo.gate = ".",
  expo = TRUE,
  expo.type = ".",
  name.dict = ".",
  foN.dict = ".",
  verbose = ".",
  dev = FALSE
)
}
\arguments{
\item{gs}{A gating set as produced by \code{\link{makeAddGatingSet}}.}

\item{expo.gate}{Which gate to export XXX.}

\item{expo}{Logical, if extracted data should exported at all.}

\item{expo.type}{Character length one. The filetype of the data export. 
Possible values are 'csv' and 'xlsx'.}

\item{name.dict}{Character length one. The name of the dictionary.}

\item{foN.dict}{Character length one. The name of the folder where the 
dictionary resides.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}

\item{dev}{Logical. If set to true, a histogram showing the bins and the
smoothed mid-points is plotted. (Only intended for development.) Defaults to
FALSE.}
}
\value{
A matrix of class "fdmat" with one row for every FCS file in the 
gating set.
}
\description{
Extract fluorescence distribution along a specified channel 
from the gating set as defined in the gating strategy file and re-calculate 
data to events per volume.
}
\examples{
\dontrun{
XXX
}
}
\seealso{
\code{\link{makeAddGatingSet}}, \code{\link{flowdexit}}

Other Extraction functions: 
\code{\link{addGates}()},
\code{\link{makeAddGatingSet}()},
\code{\link{makeGatingSet}()}
}
\concept{Extraction functions}
