% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{makeGatingSet}
\alias{makeGatingSet}
\title{Make Gating Set}
\usage{
makeGatingSet(
  patt = NULL,
  comp = ".",
  fn = ".",
  tx = ".",
  channel = ".",
  verbose = "."
)
}
\arguments{
\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.}

\item{comp}{Logical. If compensation should be applied or not. If left at 
the default '.', the value as defined in the settings file (key 'dV_comp') 
will be used.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file will be used.}

\item{tx}{Character length one. The transformation applied to *all* channels 
within the individual flow sets. If left at the default '.', the value as 
defined in the settings file (key 'dV_tx') will be used. (Currently only 
'fjbiex' is implemented.)}

\item{channel}{A regular expression indicating which channels, i.e. which
columns to keep from the original flowframes; is passed down to argument
\code{column.pattern} of \code{\link[flowCore]{read.flowSet}}. Set to NULL
to read data from all channels. If left at the default '.', the value as 
defined in the settings file (key 'dV_channel') will be used.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
A gating set produced by \code{\link[flowWorkspace]{GatingSet}}.
}
\description{
Read in FCS files and put them together in a gating set.
}
\details{
If no folder name is specified, FCS files are being read in from the 
default FCS-files folder.
}
\section{Regarding Compensation}{
 Due to the circumstances when developing 
this code, it was never required to apply any kind of compensation. The 
functionality to apply compensation was therefore never tested or verified. 
It is strongly advised to use caution when applying compensation. It might 
well be necessary to modify the source code of this function 
(fork from \url{https://github.com/bpollner/flowdex}) in order to achieve 
correct compensation results.
}

\concept{Extraction functions}
