% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{repairSID}
\alias{repairSID}
\title{Repair a single Sample ID.}
\usage{
repairSID(
  fs = NULL,
  name = NULL,
  newSID = NULL,
  patt = NULL,
  fn = ".",
  confirm = TRUE,
  ignore.text.offset = FALSE
)
}
\arguments{
\item{fs}{The object returned by this function if parameter \code{fs} is left
at its default NULL, what then can be used as input for the parameter \code{fs}.}

\item{name}{Character length one. The name of the fcs file within the flowset 
that should get a new sample ID.}

\item{newSID}{Character legnth one. The new Sample ID.}

\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file (key: 'foN_fcsFiles') will be used.}

\item{confirm}{Logical. If the user should be asked for additional confirmation
before the rewriting of the fcs file is performed. Defaults to TRUE.}

\item{ignore.text.offset}{Logical. If set to true, fcs-files in the folder 
specified at argument 'fn' will be checked for inconsistencies in the 
HEADER and TEXT segment and, if those inconsistency is present, the 
afflicted fcs-file will be re-written to disc, with the values in TEXT 
being ignored. The file will be \strong{overwritten without further warning.}}
}
\value{
Nothing, called for its side effects: the specified single fcs file gets
written to disc with its new sample ID.
}
\description{
Replace a faulty sample ID with a new one and write the single
fcs file back to disc.
}
\details{
To first obtain the flowSet, leave the parameter `fs`at its default
NULL. By providing a pattern to `patt`, subgroups of fcs files can be read in.
Provide the so obtained flowSet to the parameter `fs`, and specify the
name and the new sample ID in order to re-write the specified file with its
new Sample ID. 'object@phenoData@data' can be used to inspect and verify names 
of FCS files and the sample IDs therein -- see examples.
}
\section{Note}{
 The sample ID, esp. the correct sample ID is of importance 
when using the 'dictionary' to expand the abbreviations in the sample ID - 
see XXX.
}

\examples{
\dontrun{
fs <- repairSID() # or 
fs <- repairSID(patt="foo")
fs@phenoData@data
repairSID(fs, "sample1", "newSID")
}
}
\seealso{
Other Repair functions: 
\code{\link{repairVolumes}()}
}
\concept{Repair functions}
