% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{repairVolumes}
\alias{repairVolumes}
\title{Repair Volume Values in FCS files}
\usage{
repairVolumes(
  patt = NULL,
  vol = NULL,
  fn = ".",
  includeAll = FALSE,
  confirm = TRUE,
  fcsRepair = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.}

\item{vol}{Numeric length one. The value that should be written into the
\code{$VOL} slot in the description of the single flowFrame.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file (key: 'foN_fcsFiles') will be used.}

\item{includeAll}{Logical. If left at the default \code{FALSE}, only those
FCS files with *missing* volume information are read in and the value provided
in \code{vol} will be assigned to them. If changed to \code{TRUE}, also FCS
files with *present* volume information will be read in and the value provided
in \code{vol} will be assigned to them.}

\item{confirm}{Logical. If the user should be asked for additional confirmation
before the rewriting of the fcs files is performed. Defaults to TRUE.}

\item{fcsRepair}{Logical. If set to TRUE, fcs-files in the folder 
specified at argument 'fn' will be checked for multiplied entries in the 
keywords, as after some testing the author came to the humble conclusion that 
these multiplied keywords can be the reason for the error message: \cr
\code{"The HEADER and the TEXT segment define different starting 
point ... to read the data"}. \cr
If 'fcsRepair' is set to TRUE, all except the last of each multiplied 
keyword will be removed and the fcs file will be saved to disc, 
overwriting the original fcs file \strong{without further warning}. \cr
Use the function \code{\link{checkRepairFcsFiles}} which does offer more 
options to manually check and repair afflicted fcs files. There, it is 
possible to display multiplied keywords and to select which one to keep.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
Nothing, resp. the modified fcs files written back into the same folder
and replacing the original ones.
}
\description{
Read in all or a subset of FCS files, replace missing volume
values in the description of the single flowFrames with a provided default
value and write the FCS-files back to disc.
}
\details{
This function is intended to be used when, for known or unknown
reasons, there is no volume information encoded in the original fcs file.
}
\seealso{
Other Accessory functions: 
\code{\link{applyBandpass}()},
\code{\link{checkRepairFcsFiles}()},
\code{\link{cutFdmatToGate}()},
\code{\link{exportFdmatData}()},
\code{\link{fd_load}()},
\code{\link{fd_save}()},
\code{\link{genfs}()},
\code{\link{repairSID}()}

Other Repair functions: 
\code{\link{checkRepairFcsFiles}()},
\code{\link{repairSID}()}
}
\concept{Accessory functions}
\concept{Repair functions}
