% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{repairVolumes}
\alias{repairVolumes}
\title{Repair Volume Values in FCS files}
\usage{
repairVolumes(
  patt = NULL,
  vol = NULL,
  fn = ".",
  includeAll = FALSE,
  confirm = TRUE,
  ignore.text.offset = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{patt}{A regular expression defining a possible subset of FCS files
residing in the directory specified by \code{fn} to read in. Only matching
patterns will be included.}

\item{vol}{Numeric length one. The value that should be written into the
\code{$VOL} slot in the description of the single flowFrame.}

\item{fn}{Character length one. The name of the folder where FCS files should 
be read from. If left at the default '.', the folder name as defined in the 
settings file will be used.}

\item{includeAll}{Logical. If left at the default \code{FALSE}, only those
FCS files with *missing* volume information are read in and the value provided
in \code{vol} will be assigned to them. If changed to \code{TRUE}, also FCS
files with *present* volume information will be read in and the value provided
in \code{vol} will be assigned to them.}

\item{confirm}{Logical. If the user should be asked for additional confirmation
before the rewriting of the fcs files is performed. Defaults to TRUE.}

\item{ignore.text.offset}{Logical. If set to true, fcs-files in the folder 
specified at argument 'fn' will be checked for inconsistencies in the 
HEADER and TEXT segment and, if those inconsistency is present, the 
afflicted fcs-file will be \strong{re-written to disc without further warning}, 
with the values in TEXT being ignored.}

\item{verbose}{Logical. If status messages should be displayed. If left at 
the default '.', the value as defined in the settings file (key 'dV_verbose') 
will be used.}
}
\value{
Nothing, resp. the modified fcs files written back into the same folder
and replacing the original ones.
}
\description{
Read in all or a subset of FCS files, replace missing volume
values in the description of the single flowFrames with a provided default
value and write the FCS-files back to disc.
}
\details{
This function is intended to be used when, for known or unknown
reasons, there is no volume information encoded in the original fcs file.
}
\seealso{
Other Repair functions: 
\code{\link{repairSID}()}
}
\concept{Repair functions}
