% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen-functions.R
\name{setupSettings}
\alias{setupSettings}
\title{Perform settings setup}
\usage{
setupSettings(path = NULL)
}
\arguments{
\item{path}{Character length one, holding the path to the location where the 
folder containing the settings.R file should be located. Defaults to 'NULL'. 
If left at the default 'NULL', the location should be selectable interactively.}
}
\value{
No return value, is called for its side effects, i.e. to set up 
the settings-file system as provided by package 'uniset'.
}
\description{
Perform the setup of the setting-system provided by package 
'uniset'.
}
\details{
Has to be done only once.
}
\examples{
\dontrun{
setupSettings(path="~/foo/bar")
}
}
\seealso{
\code{\link{fd_updateSettings}}
}
